MSD-5 Enables multi-platform builds
Adds support for building images on both linux/amd64 and linux/arm64 architectures.

This change optimizes the build process by configuring Buildx to build for multiple platforms, and sets the platforms in the build and push steps to ensure images are built for both amd64 and arm64 architectures.
This enables wider compatibility for the built images.