MSD-5 Adds build and push script for ARM64
Introduces a new script to build and push Docker images for ARM64 architecture.

Also adds a Dockerfile specifically tailored for ARM64 builds. This allows for containerized applications to be built and deployed on ARM64 platforms.
The script automates the process of building, tagging, and pushing the image to GitHub Container Registry (ghcr.io), using the git history for versioning.

Ignores local configuration files to prevent sensitive information from being committed.